requirements.txt
pydantic-ai
python-dotenv
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0